1.use namespace user_map in user_map module 2.get mac/x/y/offset from kafka message to call user_map::user_add 3.fix the bug of initialize tair-rdb client

+++ ad_model.h
@@ -70,7 +70,7 @@ struct AdData
 
 struct UserPosition
 {
-    int user_id;
+    unsigned long long user_id;
     GeoInfo position;
     int valid;
 };
